Time: 1 ms (97.44%), Space: 42.8 MB (23.51%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -1,7 +1,7 @@
 class Solution {
     public int mySqrt(int x) {
-        long sqrt = 0;
-        sqrt = (long) Math.sqrt(x);
-        return (int)sqrt;
+        int sqrt = 0;
+        sqrt = (int) Math.sqrt(x);
+        return sqrt;
     }
 }
